cong tuner needs to set congestion control for existing connections too
when we see retransmits to a remote host, we'd like to switch on bbr
for all connections, not just new ones.  The only workable approach
that allows us to change congestion control algorithm for each socket
is via bpf iter. We attach an iter to tcp sockets, and when we get
enough retransmits for a remote host we send an event which triggers
iterator traversal, using the map of remote host IPs as a basis for
determining which connections to set bbr for.