add correlation map + correlation computations to bpftune
we want to share a correlation map where we can correlate tunable + variable
of interest; currently keyed by tunable + netns cookie but many need more
key fields to support multiple metrics to correlate for a tunable.

currently we correlate [wr]mem increases and srtt_us (smoothed round-trip
time) to spot cases where increases to buffer size are correlated with
longer latencies. currently we do not do anything with this info but shortly
we will drive buffer values back down such that the correlation is below
0.5; this is a pull back to combat the push of increasing mem limits
forever.

in practice for high wmem values we see strong correlation > 0.8 with
srtt, but for the most part rmem increases are uncorrelated.