bpftune: legacy support for systems that do not have fentry etc
For older systems that do not have fentry/fexit, we need a fallback;
we fallback to using kprobes in such cases.  The first step is
providing wrapper definitions for BPF progs such that we can
create a legacy object/skeleton to fall back to if attach to
the main object fails.  Still some work to be done

- handle fexit->kretprobe
- handle cases where we use new features like iterators;
  replace with tcpbpf instead guarded with #ifdef BPFTUNE_LEGACY
- document legacy support
- add tests using force legacy option