fix legacy issues, add legacy support for netns tuner
netns tuner uses fexit so needs access to entry args on return;
in legacy use kprobe/kretprobe combo with map to store entry
netns so we can use it on return.

a few issues with legacy skel code; we were not using lskel in
a few places, so hit issues with dereferencing the null _skel.

add netns legacy tuner test