add inotify support to notice (dis)appearance of tuners
we want to support separate package deliver of tuners, so by
using inotify watching /usr/lib64/bpftune we can catch these
events.  One complication is that we need to be careful around
fd reuse - if the original skeleton that sets ringbuf fd
for reuse is destroyed, the ringbuf map fd becomes invalid.

So we pin common map fds under /sys/fs/bpf/bpftune and
dup the associated map fds for reuse; this avoids the issue
when a tuner disappers/later reappears.