reading /proc/kallsyms requires CAP_SYSLOG
without it libbpf cannot populate kernel symbol __ksym addresses
and tests that rely on them fail. in particular we get init_net
so that we can spot the global netns in the absence of netns
cookies, so the sysctl test fails since it cannot match
namespace associated with a sysctl with the global netns
(since init_net is 0 as kallsyms emits all-zero addresses
when capabilities are not present).

this bug only seen for 5.4 kernels which do not have netns
cookies as we rely on them for newer kernels; introduced by
capability-awareness code.