add support for optional tunables
sometimes tunables may not be present or accessible, e.g. when

net.ipv6.conf.all.disable_ipv6 = 1

mark such tunables as optional so tuner will not fail to load
if for example v6 tunables are not available.  Also add test
for this (neigh_table_v4only_test) which disables IPv6 while
verifying the v4 aspects of the tuner still work.