add multiple strategy support
tuners can support multiple strategies; strategies should be
set in the init() method via bpftune_strategies_add().  A
strategy specifies a
 - name
 - description
 - timeout
 - evaluation function
 - set of bpf program names in tuner associated with strategy

The current strategy runs until timeout, then the evaluation
functions for all strategies are called; we then switch to
the strategy with the highest evaluation.  See tests/strategy
for an example of a tuner implementing multiple strategies.