bpftune: support a "no BTF" mode of operation
on some systems BPF Type Format (BTF) info is not available.  Because
we make extensive use of it, we need a separate set of BPF programs that
work in a "no BTF" world; these cannot make use of Compile Once - Run
Everywhere (CO-RE) relocations, so data structure offsets may be wrong
for the running kernel, so warn the user.

Tuners cannot be supported in this mode because the likeliehood of such
issues is nearly 100%; warn the user about this.

Reported-by: Stuart Shelton <https://github.com/srcshelton>
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>