use bpf global for netdev_max_backlog as ksym changes in 6.9
net_buffer_tuner.so was failing to load BTF due to a failure
to find netdev_max_backlog ksym.

netdev_max_backlog was an int prior to 6.9 then it got added
to a netdev hotpath struct; rather than using it in calculations
retrieve the sysctl value and set a bpf global for use in bpf
prog context; update it in userspace when we bump sysctl value.

Reported-by: HippieMitch <https://github.com/HippieMitch>
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>