libbpftune: support quiet mode for bpftune_netns_set()
when matching netns cookies with netns, we iterate over all
netns and setns() to retrieve the cookie. This can fail but
currently we dump error messages in such cases. Error messaging
should be reserved for the case where we are entering the netns
to either read or write sysctls.  So add a bool quiet param
to bpftune_netns_set().

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Reported-by: Rob Landers <landers.robert@gmail.com>