libbpftune: only send events to active tuners
Roger reported a segmentation fault in tcp_buffer_tuner. Examining the
state of the tuner we were handling an event when in manual (i.e. off)
state.  The event handler for a tuner should only be called if the tuner
is active.

Reported by: Roger Knobbe (https://github.com/rknobbe)
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>