libbpftune: only show errors on optional prog reload
when optional programs are included in a tuner and the initial
(load all programs) stage fails, we fall back to loading only
the non-optional progs.  However we still get error messages
from the first case, which makes it look like the tuner ultimately
failed to load.  Add a quiet option for tuner load in the case
of initial load all programs phase where we have optional programs
for later retry.

Reported-by: NullX47 (https://github.com/NullX47)
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>