Added changes to try reduce the false positives
If you only try detection based on network traffic you get spikes based
on other apps like systemd-resolve, torrents, steam itself, browser,
discord...etc. I added a blacklist and a whitelist, the blacklist is for
common stuff I know isn't gaming related, like utp::*. For gaming
workloads I assume that it is proton, or from a launcher or whatever. It
has a corner case that if a game is natively running on the system you
would need to use gamemoderun or game-performance to hint to the
userspace side of the tool but it is a corner case and I'm sure someone
smarter than me can do it better.