neigh_table_tuner: compute/enforce upper limit to gc_thresh3
The neigh_table_tuner uses netlink to update per-neighbour table
values for the specified device.  The problem is we do not have
an upper limit on how much we update.  Here we compute a heuristic
based upon the prefixlens of the IP addresses configured for that
device; this gives us a ballpark figure for the number of addresses
possible on that network.  We then multiply that by a scaling factor
to allow for multicast etc.

For the simple case of a single /24 IPv4 address configured for the
device, we compute

256 * NEIGH_TABLE_ADDR_FACTOR (4) = 1024

This becomes the upper bound.  If more addresses/larger networks
are configured, it will be adjusted.

Also update docs and tests, adding neigh_table_max_test which
exhausts max size for ARP table via excessive multicast addresses,
and ensures updates stop at limit.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>