net_buffer_tuner: set high_order_alloc_disable=0 on newer kernels
The documentation for net.core.high_order_alloc_disable says

"By default the allocator for page frags tries to use high order pages (order-3
on x86). While the default behavior gives good results in most cases, some users
might have hit a contention in page allocations/freeing. This was especially
true on older kernels (< 5.14) when high-order pages were not stored on per-cpu
lists. This allows to opt-in for order-0 allocation instead but is now mostly of
historical importance."

So we set high_order_alloc_disable=0 on tuner init for kernels >= 5.14.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>